requirements.txt
python-telegram-bot==20.7
aiomysql==0.2.0
cryptography==42.0.5
python-dotenv==1.0.1
tronpy==0.5.0
qrcode==7.4.2
Pillow==10.3.0
pydantic==2.6.4
